fix clippy no default features
